--- features/ajna/common/types/AjnaAction.types.ts
@@ -1,22 +1,0 @@
-export type AjnaBorrowAction =
-  | 'open-borrow'
-  | 'deposit-borrow'
-  | 'withdraw-borrow'
-  | 'generate-borrow'
-  | 'payback-borrow'
-  | 'switch-borrow'
-  | 'close-borrow'
-  | 'adjust-borrow'
-
-export type AjnaEarnAction = 'open-earn' | 'deposit-earn' | 'withdraw-earn' | 'claim-earn'
-
-export type AjnaMultiplyAction =
-  | 'open-multiply'
-  | 'adjust'
-  | 'deposit-collateral-multiply'
-  | 'deposit-quote-multiply'
-  | 'generate-multiply'
-  | 'payback-multiply'
-  | 'withdraw-multiply'
-  | 'switch-multiply'
-  | 'close-multiply'
