+++ features/ajna/positions/common/components/AjnaDupePositionModal.tsx
@@ -2,8 +2,8 @@ import { NetworkIds } from 'blockchain/networks'
 import type { UserDpmAccount } from 'blockchain/userDpmProxies.types'
 import { AppLink } from 'components/Links'
 import { Modal, ModalCloseIcon } from 'components/Modal'
-import type { ProtocolProduct } from 'features/ajna/common/types'
 import { getOraclessProductUrl } from 'features/poolFinder/helpers'
+import type { ProtocolProduct } from 'features/unifiedProtocol/types'
 import { useModalContext } from 'helpers/modalHook'
 import { staticFilesRuntimeUrl } from 'helpers/staticPaths'
 import { startCase } from 'lodash'
