+++ features/ajna/positions/common/getAjnaSidebarTitle.tsx
@@ -1,5 +1,5 @@
 import type { AjnaEarnPosition, AjnaPosition } from '@oasisdex/dma-library'
-import type { ProtocolProduct, ProtocolSidebarStep } from 'features/ajna/common/types'
+import type { ProtocolProduct, ProtocolSidebarStep } from 'features/unifiedProtocol/types'
 import { zero } from 'helpers/zero'
 import { upperFirst } from 'lodash'
 import { useTranslation } from 'next-i18next'
