+++ features/ajna/positions/common/helpers/getFlowStateConfig.ts
@@ -1,7 +1,7 @@
 import type BigNumber from 'bignumber.js'
-import type { ProtocolFlow } from 'features/ajna/common/types'
 import type { AjnaFormState } from 'features/ajna/common/types/AjnaFormState.types'
 import { getMaxIncreasedValue } from 'features/ajna/positions/common/helpers/getMaxIncreasedValue'
+import type { ProtocolFlow } from 'features/unifiedProtocol/types'
 import type { UseFlowStateProps } from 'helpers/useFlowState'
 import { zero } from 'helpers/zero'
