+++ features/ajna/positions/common/hooks/useAjnaData.ts
@@ -6,9 +6,9 @@ import { useAccountContext } from 'components/context/AccountContextProvider'
 import { useMainContext } from 'components/context/MainContextProvider'
 import { useProductContext } from 'components/context/ProductContextProvider'
 import { isPoolOracless } from 'features/ajna/common/helpers/isOracless'
-import type { ProtocolProduct } from 'features/ajna/common/types'
 import { getAjnaPositionAggregatedData$ } from 'features/ajna/positions/common/observables/getAjnaPositionAggregatedData'
 import { getStaticDpmPositionData$ } from 'features/ajna/positions/common/observables/getDpmPositionData'
+import type { ProtocolProduct } from 'features/unifiedProtocol/types'
 import { getPositionIdentity } from 'helpers/getPositionIdentity'
 import { useObservable } from 'helpers/observableHook'
 import { useAccount } from 'helpers/useAccount'
