+++ features/ajna/positions/common/observables/getAjnaPosition.ts
@@ -7,13 +7,14 @@ import type { Tickers } from 'blockchain/prices.types'
 import type { UserDpmAccount } from 'blockchain/userDpmProxies.types'
 import type { PositionCreated } from 'features/aave/services'
 import { isPoolOracless } from 'features/ajna/common/helpers/isOracless'
-import type { AjnaGenericPosition, ProtocolProduct } from 'features/ajna/common/types'
+import type { AjnaGenericPosition } from 'features/ajna/common/types'
 import { getAjnaCumulatives } from 'features/ajna/positions/common/helpers/getAjnaCumulatives'
 import { getAjnaPoolAddress } from 'features/ajna/positions/common/helpers/getAjnaPoolAddress'
 import { getAjnaPoolData } from 'features/ajna/positions/common/helpers/getAjnaPoolData'
 import type { DpmPositionData } from 'features/ajna/positions/common/observables/getDpmPositionData'
 import { getAjnaEarnData } from 'features/ajna/positions/earn/helpers/getAjnaEarnData'
 import { checkMultipleVaultsFromApi$ } from 'features/shared/vaultApi'
+import type { ProtocolProduct } from 'features/unifiedProtocol/types'
 import { one } from 'helpers/zero'
 import { LendingProtocol } from 'lendingProtocols'
 import { isEqual, uniq } from 'lodash'
