+++ features/ajna/positions/common/observables/getAjnaPositionAggregatedData.ts
@@ -1,13 +1,14 @@
 import type { AjnaEarnPosition, AjnaPosition } from '@oasisdex/dma-library'
 import type { NetworkIds } from 'blockchain/networks'
 import dayjs from 'dayjs'
-import type { AjnaGenericPosition, ProtocolProduct } from 'features/ajna/common/types'
+import type { AjnaGenericPosition } from 'features/ajna/common/types'
 import type { AjnaUnifiedHistoryEvent } from 'features/ajna/history/ajnaUnifiedHistoryEvent'
 import type { AjnaPositionAggregatedDataAuctions } from 'features/ajna/positions/common/helpers/getAjnaPositionAggregatedData'
 import { getAjnaPositionAggregatedData } from 'features/ajna/positions/common/helpers/getAjnaPositionAggregatedData'
 import { mapAjnaBorrowishEvents } from 'features/ajna/positions/common/helpers/mapBorrowishEvents'
 import { mapAjnaEarnEvents } from 'features/ajna/positions/common/helpers/mapEarnEvents'
 import type { DpmPositionData } from 'features/ajna/positions/common/observables/getDpmPositionData'
+import type { ProtocolProduct } from 'features/unifiedProtocol/types'
 import { zero } from 'helpers/zero'
 import type { Observable } from 'rxjs'
 import { from } from 'rxjs'
