+++ features/ajna/positions/earn/helpers/getAjnaEarnDefaultAction.ts
@@ -1,5 +1,5 @@
 import type { AjnaEarnPosition } from '@oasisdex/dma-library'
-import type { ProtocolFlow } from 'features/ajna/common/types'
+import type { ProtocolFlow } from 'features/unifiedProtocol/types'
 import { zero } from 'helpers/zero'
 
 export const getAjnaEarnDefaultAction = (flow: ProtocolFlow, ajnaPosition: AjnaEarnPosition) =>
