+++ pages/[networkOrProduct]/ajna/[...position].tsx
@@ -5,8 +5,8 @@ import { ProductContextHandler } from 'components/context/ProductContextHandler'
 import { isAddress } from 'ethers/lib/utils'
 import { ajnaProducts } from 'features/ajna/common/consts'
 import { AjnaLayout, ajnaPageSeoTags } from 'features/ajna/common/layout'
-import type { ProtocolProduct } from 'features/ajna/common/types'
 import { AjnaProductController } from 'features/ajna/positions/common/controls/AjnaProductController'
+import type { ProtocolProduct } from 'features/unifiedProtocol/types'
 import type { GetServerSidePropsContext } from 'next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import React from 'react'
